pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Victor Banda <victor.banda@gosh.nhs.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.16"
poetry = "1.3.2"
flashtext = "2.7"
beautifulsoup4 = "4.12.2"
ipywidgets = "8.0.6"
lxml = "4.9.3"
nltk = "3.9.1"
numpy = "1.24.4"
pandas = "2.0.3"
pickleshare = "0.7.5"
pip = "23.0.1"
preshed = "3.0.8"
presidio-analyzer = "2.2.33"
presidio-anonymizer = "2.2.33"
pysbd = "0.3.4"
pyodbc = "4.0.39"
PyYAML = "6.0"
regex = "2023.3.23"
requests = "2.31.0"
scikit-learn = "1.0.2"
SQLAlchemy = "1.4.46"
tqdm = "4.64.1"
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1.tar.gz"}
pytest = "^7.4.3"
emoji = "^2.12.1"
setfit = "^1.0.3"
huggingface-hub = "0.23.5"
transformers-interpret = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
